Reduce lifetime of static assets (#102)
Most of this app is static assets and it seems to me that there is no cache busting implemented for these assets. Therefore I suggest we set the default cache lifetime so low that it facilitates publishing new versions of the application.

When I request the home page, separate requests are done for all the different bower files etc. If there is a way they should be bundled, I have not found it. This was the fix I implemented :)

Regards,
Tarjei