Add NSIS differential update support
The NSIS differential update support is tested with electron-builder (https://www.electron.build/)

The AssetService.getHash function is rewritten to match the hash check
of electron-builder:
https://github.com/electron-userland/electron-builder/blob/552f1a4ed6f4bb83c3c548ed962c21142f07a9b4/packages/electron-updater/src/DownloadedUpdateHelper.ts#L161
They use base64 encoding for storing the hash (instead of hex), so now
we have an extra parameter to controll the encoding.

This paramter is used for calculating the hash of a .exe or .zip file.
We also switch to sha512 because the sha256 check does not work correctly
for the differential update support of electron-builder.

Also, the latest.yml (or beta.yml / alpha.yml) now matches the .yml files
that are generated by electron-builder.
This is done for Windows and Mac .yml files.
However, differential updates do not work for MacOS because electron-builder
doesn't properly support it.

For the differential updates to work, you have to upload a blockmap.
So now the .blockmap files are also accepted as uploaded files.