Fixing bug with bootstrap-sass. (#291)
Newer versions of bootstrap-sass-official are not compatible with the version of sass used in this project as it includes changes to the way division works. This issue is documented in these issues; https://github.com/twbs/bootstrap-sass/issues/1231 and https://github.com/twbs/bootstrap-sass/issues/1230. Making the version of bootstrap-sass required more specific stops the breaking upgrade.